requirements.txt
# WayYaSnitch - CPU Only Dependencies
# No GPU acceleration required

PyQt5>=5.15.0
opencv-python-headless>=4.5.0
numpy>=1.20.0
Pillow>=8.0.0
img2pdf>=0.4.0
pynput>=1.7.0
